move timeout functions and classes to util.timeout
