+++ paasta_tools/cli/cmds/local_run.py
@@ -41,6 +41,8 @@ from paasta_tools.long_running_service_tools import get_healthcheck_for_instance
 from paasta_tools.paasta_execute_docker_command import execute_in_container
 from paasta_tools.secret_tools import decrypt_secret_environment_variables
 from paasta_tools.tron_tools import parse_time_variables
+from paasta_tools.util.timeout import Timeout
+from paasta_tools.util.timeout import TimeoutError
 from paasta_tools.utils import _run
 from paasta_tools.utils import DEFAULT_SOA_DIR
 from paasta_tools.utils import get_docker_client
@@ -56,8 +58,6 @@ from paasta_tools.utils import PaastaColors
 from paasta_tools.utils import PaastaNotConfiguredError
 from paasta_tools.utils import SystemPaastaConfig
 from paasta_tools.utils import timed_flock
-from paasta_tools.utils import Timeout
-from paasta_tools.utils import TimeoutError
 from paasta_tools.utils import validate_service_instance
