+++ paasta_tools/mesos_tools.py
@@ -53,6 +53,7 @@ from paasta_tools.mesos.exceptions import SlaveDoesNotExist
 from paasta_tools.mesos.master import MesosMaster
 from paasta_tools.mesos.master import MesosState
 from paasta_tools.mesos.task import Task
+from paasta_tools.util.timeout import TimeoutError
 from paasta_tools.utils import DeployBlacklist
 from paasta_tools.utils import DeployWhitelist
 from paasta_tools.utils import format_table
@@ -60,7 +61,6 @@ from paasta_tools.utils import get_user_agent
 from paasta_tools.utils import load_system_paasta_config
 from paasta_tools.utils import PaastaColors
 from paasta_tools.utils import SystemPaastaConfig
-from paasta_tools.utils import TimeoutError
 
 MARATHON_FRAMEWORK_NAME_PREFIX = "marathon"
